+++ src/frontend/components/TrackViewerFrame.tsx
@@ -7,6 +7,5 @@ import { MAP_VIEWER_HEIGHT } from '../constants/map-viewer';
  * @see https://developer.atlassian.com/platform/forge/ui-kit/components/frame/
  */
 export const TrackViewerFrame = (): React.JSX.Element => (
-  // @ts-expect-error UI Kit Frame supports height/width; @forge/react typings omit them
   <Frame resource="track-viewer" height={MAP_VIEWER_HEIGHT} width="100%" />
 );